Made a lot of improvements on the Read Generator.php of the story websites

+++ Variables/Global Files/Variable Inserter.php
@@ -23,31 +23,89 @@ function Variable_Name($var) {
 function Variable_Inserter($array, $text_line) {
 	global $verbose;
 	global $has_variable;
+	global $variable_inserter_replacer_array;
+	global $use_text_as_replacer;
 
 	$backup_text_line = $text_line;
 
 	$has_variable = False;
 
-	foreach ($array as $variable) {
-		$variable_name = Variable_Name($variable);	
+	if ($use_text_as_replacer == True) {
+		$i = 0;
+		foreach ($variable_inserter_replacer_array as $text) {
+			$variable = $array[$i];
 
-		if ($verbose == True) {
-			echo "<br />Linha: ".$text_line."<br />"."Nome da Variável: $".$variable_name.";<br />"."Variável: $".$variable.";<br />";
+			if ($verbose == True) {
+				echo "<br />Linha: ".$text_line."<br />"."Variável: ".$variable."<br />"."Texto: ".$text."<br />";
+			}
+
+			if (strpos($text_line, $text) == True and $has_variable == False) {
+				$new_text_line = $text_line;
+
+				$new_text_line = str_replace($text, $variable, $new_text_line);
+
+				$has_variable = True;
+			}
+
+			if (strpos($text_line, $text) == False and $has_variable == False) {
+				$new_text_line = $text_line;
+
+				$new_text_line = str_replace(array(".", "?", ")", "(", ":", "/", "="), "", $text_line);
+
+				echo "<br />Linha: <br />".$new_text_line."<br />"."Texto: <br />".str_replace(array(".", "?", ")", "(", ":", "/", "="), "", $text)."<br /><br />";
+
+				if ($new_text_line == str_replace(array(".", "?", ")", "(", ":", "/", "="), "", $text)) {
+					$new_text_line = str_replace($text, $variable, $new_text_line);
+
+					$has_variable = True;
+				}
+			}
+
+			if ($has_variable == True) {
+				$text_line = $new_text_line;
+			}
+
+			if ($has_variable == False) {
+				$text_line = $backup_text_line;
+			}
+
+			$i++;
 		}
+		
+	}
 
-		if (strpos($text_line, "$".$variable_name.";") == True and $has_variable == False) {
-			$new_text_line = str_replace("$".$variable_name.";", $variable, $text_line);
+	if ($use_text_as_replacer == False) {
+		foreach ($array as $variable) {
+			$variable_name = Variable_Name($variable);	
 
-			$has_variable = True;
-		}
+			if ($verbose == True) {
+				echo "<br />Linha: ".$text_line."<br />"."Nome da Variável: $".$variable_name.";<br />"."Variável: $".$variable.";<br />";
+			}
 
-		if (strpos($text_line, "$".$variable_name.";") == False and $has_variable == False) {
-			$new_text_line = str_replace(array(".", "?", ")", "(", " SpaceLiving Network"), "", $text_line);
+			if (strpos($text_line, "$".$variable_name.";") == True and $has_variable == False) {
+				$new_text_line = $text_line;
 
-			if ($new_text_line == "$".$variable_name.";") {
+				if (isset($variable_inserter_replacer_array)) {
+					$new_text_line = str_replace($variable_inserter_replacer_array, "", $new_text_line);
+				}
+
 				$new_text_line = str_replace("$".$variable_name.";", $variable, $new_text_line);
 
 				$has_variable = True;
+			}
+
+			if (strpos($text_line, "$".$variable_name.";") == False and $has_variable == False) {
+				$new_text_line = str_replace(array(".", "?", ")", "(", " SpaceLiving Network"), "", $text_line);
+
+				if (isset($variable_inserter_replacer_array)) {
+					$new_text_line = str_replace($variable_inserter_replacer_array, "", $new_text_line);
+				}
+
+				if ($new_text_line == "$".$variable_name.";") {
+					$new_text_line = str_replace("$".$variable_name.";", $variable, $new_text_line);
+
+					$has_variable = True;
+				}
 			}
 		}
 	}
